Move StaticHiddenInput code to HiddenInput
