--- src/signals/incident/components/form/StaticHiddenInput/index.js
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MPL-2.0
-// Copyright (C) 2018 - 2021 Gemeente Amsterdam
-
-import StaticHiddenInput from './StaticHiddenInput'
-
-export default StaticHiddenInput
